requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
attrs==25.4.0
banks==2.2.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
blis==1.3.3
boto3==1.42.55
botocore==1.42.55
browserforge==1.2.3
camoufox==0.4.11
catalogue==2.0.10
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
cryptography==46.0.5
cymem==2.0.13
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
emoji==2.15.0
fastapi==0.132.0
filelock==3.20.3
filetype==1.2.0
Flask==3.1.3
fpdf==1.7.2
frozenlist==1.8.0
fsspec==2026.1.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.3.0
griffe==1.15.0
grpcio==1.78.1
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.3.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
installer==0.7.0
invoke==2.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.1.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.2.15
langchain-text-splitters==1.1.1
langdetect==1.0.9
langsmith==0.7.6
language-tags==1.2.0
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.12
llama-index-cli==0.5.3
llama-index-core==0.14.12
llama-index-embeddings-huggingface==0.6.1
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.2
llama-index-llms-groq==0.4.1
llama-index-llms-openai==0.6.12
llama-index-llms-openai-like==0.5.3
llama-index-readers-file==0.5.6
llama-index-readers-llama-parse==0.5.1
llama-index-workflows==2.11.7
llama-parse==0.6.54
llvmlite==0.46.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
murmurhash==1.0.15
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.2
numba==0.64.0
numpy==2.4.0
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
olefile==0.47
openai==2.15.0
opencv-python-headless==4.13.0.92
orjson==3.11.5
packaging==25.0
pandas==2.3.3
paramiko==4.0.0
pdf2image==1.17.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.0
platformdirs==4.5.1
playwright==1.57.0
preshed==3.0.12
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.2
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.11.0
PyMuPDF==1.27.1
PyNaCl==1.6.2
pyparsing==3.3.2
pypdf==6.6.0
pypdfium2==5.5.0
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-iso639==2026.1.31
python-magic==0.4.27
python-multipart==0.0.22
python-oxmsg==0.0.2
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
s3transfer==0.16.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
screeninfo==0.8.1
sentence-transformers==5.2.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
sniffio==1.3.1
soupsieve==2.8.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.45
srsly==2.5.2
starlette==0.52.1
striprtf==0.0.26
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
thinc==8.3.10
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.9.1
tqdm==4.67.3
transformers==4.57.3
triton==3.5.1
typer==0.24.1
typer-slim==0.24.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ua-parser==1.0.1
ua-parser-builtins==202601
unstructured==0.21.2
unstructured-client==0.42.10
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
wasabi==1.1.3
weasel==0.4.3
webencodings==0.5.1
Werkzeug==3.1.6
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
